+++ spec/test_factories/car_factory.rb
@@ -23,8 +23,5 @@ Satisfactory::UnitFactories.configurations[:car] = {
     plural: [:wheels],
     singular: [:engine],
   },
-  model: Car,
-  name: :car,
   parent: nil,
-  traits: [],
 }